Refactor: Fix minor issues and improve type safety across components
