+++ src/components/task/TaskCommentSection.tsx
@@ -282,9 +282,10 @@ export const TaskCommentSection: React.FC<TaskCommentSectionProps> = ({
       let match;
 
       while ((match = mentionRegex.exec(commentText)) !== null) {
-        const _mentionedName = match[1];
+        const mentionedName = match[1];
         // 그룹 멤버에서 이름으로 ID 찾기
         const mentionedUser = groupMembers?.find(
+          member => member.displayName === mentionedName
         );
         if (mentionedUser) {
           mentions.push(mentionedUser.userId);
