+++ src/contexts/AppContextDef.tsx
@@ -1,4 +1,4 @@
 import { createContext } from 'react';
-import { AppContextType } from './AppContext';
+import { AppContextType } from './AppContextTypes';
 
 export const AppContext = createContext<AppContextType | undefined>(undefined);
