+++ src/contexts/AuthContext.tsx
@@ -11,7 +11,7 @@ import {
   signInWithPopup,
   updateProfile,
 } from 'firebase/auth';
-import { serverTimestamp } from 'firebase/firestore';
+import { serverTimestamp, Timestamp } from 'firebase/firestore';
 import React, { useEffect, useState } from 'react';
 
 import { auth } from '../lib/firebase';
@@ -93,10 +93,10 @@ export function AuthProvider({ children }: { children: React.ReactNode }) {
         // Update login count and last login with safe data
         const updateData: Partial<User> = {
           loginCount: ((existingProfile as { loginCount?: number }).loginCount || 0) + 1,
-          lastLoginAt: serverTimestamp(),
-          lastLoginTime: serverTimestamp(), // 추가: 최근 로그인 시간
-          lastSignInTime: serverTimestamp(), // 추가: 최근 로그인 시간 (다른 필드명)
-          updatedAt: serverTimestamp(), // 추가: 업데이트 시간
+          lastLoginAt: serverTimestamp() as Timestamp,
+          lastLoginTime: serverTimestamp() as Timestamp, // 추가: 최근 로그인 시간
+          lastSignInTime: serverTimestamp() as Timestamp, // 추가: 최근 로그인 시간 (다른 필드명)
+          updatedAt: serverTimestamp() as Timestamp, // 추가: 업데이트 시간
           // Only update photoURL if it's not undefined/null
           ...(user.photoURL && { photoURL: user.photoURL }),
         };
